+++ tests/testthat/test-coalesce.R
@@ -47,3 +47,9 @@ test_that("coalesce() supports data frames (#5326)", {
 
   expect_error(coalesce(as.matrix(mtcars), as.matrix(mtcars)), "matrices")
 })
+
+test_that("coalesce() supports one-dimensional arrays (#5557)", {
+  x <- array(1:10)
+  out <- coalesce(x, 0)
+  expect_equal(out, x)
+})