+++ NEWS.md
@@ -1,5 +1,7 @@
 # dplyr (development version)
 
+*  `coalesce()` accepts 1-D arrays (#5557).
+
 # dplyr 1.0.7
 
 * `across()` uses the formula environment when inlining them (#5886).